pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["hokim84 <hodk84@naver.com>"]
readme = "README.md"
packages = [{include = "airbnb_clone_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"
strawberry-graphql = "^0.152.0"
strawberry-graphql-django = "^0.8.2"
pyjwt = "^2.6.0"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
